CAMEL-21183: BrowseEndpoint should have limit/filter to optimize returned data from component implementations

+++ core/camel-management/src/test/java/org/apache/camel/management/ManagedMulticastTest.java
@@ -46,11 +46,11 @@ public class ManagedMulticastTest extends ManagementTestSupport {
         MBeanServer mbeanServer = getMBeanServer();
 
         ObjectName name = getCamelObjectName(TYPE_ENDPOINT, "mock://a");
-        Long queueSize = (Long) mbeanServer.invoke(name, "queueSize", null, null);
+        Integer queueSize = (Integer) mbeanServer.invoke(name, "queueSize", null, null);
         assertEquals(3, queueSize.intValue());
 
         name = getCamelObjectName(TYPE_ENDPOINT, "mock://b");
-        queueSize = (Long) mbeanServer.invoke(name, "queueSize", null, null);
+        queueSize = (Integer) mbeanServer.invoke(name, "queueSize", null, null);
         assertEquals(3, queueSize.intValue());
 
         name = getCamelObjectName(TYPE_PROCESSOR, "myMulticast");
